refactor: move Search from Header to separate component
